\name{surfaceDemo}
\docType{data}
\alias{surfaceDemo}
\title{
Tree and Data for Demonstrating SURFACE
}
\description{
This simulated tree and data set can be used to demonstrate the functionality of SURFACE. The vignette `surface_tutorial' demonstrates the use of the various functions included in the package using \code{surfaceDemo}
}
\usage{demodata}
\format{An object containing a tree in \code{phylo} format (\code{surfaceDemo$tree})and a list (\code{surfaceDemo$sim} containing data (\code{surfaceDemo$sim$data}) as well as features of the generating model including the true Hansen model (\code{surfaceDemo$sim$fit})
}
\source{simulated data}
\references{
Ingram, T. & Mahler, D.L. in revision, Methods in Ecology and Evolution. 
}
\keyword{datasets}
