\name{surfaceDemo}
\docType{data}
\alias{surfaceDemo}
\title{
Tree and Data for Demonstrating SURFACE
}
\description{
This simulated tree and data set can be used to demonstrate the functionality of SURFACE. The vignette `surface_tutorial' demonstrates the use of the various functions included in the package using \code{surfaceDemo}
}
\usage{surfaceDemo}
\format{A list containing a tree in \code{phylo} format (\code{surfaceDemo$tree}), and a list \code{surfaceDemo$sim}, which contains trait data (\code{surfaceDemo$sim$data}) and the other features output by \code{surfaceSimulate}, including the generating Hansen model (\code{surfaceDemo$sim$fit})
}
\source{simulated data}
\references{
Ingram, T. & Mahler, D.L. (2013) SURFACE: detecting convergent evolution from comparative data by fitting Ornstein-Uhlenbeck models with stepwise AIC. Methods in Ecology and Evolution, in press. 
}
\keyword{datasets}
